task-3/backend/requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.4
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
djongo==1.3.6
environ==1.0
idna==3.3
PyJWT==1.7.1
pymongo==3.12.1
pytz==2022.1
requests==2.27.1
sqlparse==0.2.4
urllib3==1.26.9
